Fixed more database types
Former-commit-id: cfef1af8f496d365ef16350a7569f4107245879f

+++ src/game/InstanceSaveMgr.cpp
@@ -422,7 +422,7 @@ void InstanceSaveManager::LoadResetTimes()
             }
 
             // update the reset time if the hour in the configs changes
-            uint64 oldresettime = fields[1].GetUInt64();
+            uint64 oldresettime = fields[1].GetUInt32();
             uint64 newresettime = (oldresettime / DAY) * DAY + diff;
             if(oldresettime != newresettime)
                 CharacterDatabase.DirectPExecute("UPDATE instance_reset SET resettime = '" UI64FMTD "' WHERE mapid = '%u'", newresettime, mapid);
